pyproject.toml
[project]
name = "amr-gnn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.7"
dependencies = [
    "captum>=0.8.0",
    "einops==0.7.0",
    "hydra-core>=1.3.2",
    "lightning==2.2.0",
    "pandas==2.2.0",
    "scikit-learn==1.4.0",
    "seaborn==0.13.2",
    "torch-geometric==2.4.0",
]
